Remove duplicated `pk` property from `BackendModelEntity` (#4310)
Define `BackendNode` properly as an abstract class by defining the
metaclass to be `abc.ABCMeta`. This will guarantee that if abstract
methods are added, they are also added in the concrete classes, because
classes with unimplemented abstract methods cannot be instantiated.

Also remove the duplicated `pk` properties from the `SqlaModelEntity`
and `DjangoModelEntity` classes since it is already implemented in their
base class `BackendEntity`.