`verdi computer test`: fix bug in spurious output test
The test that checks for spurious output when executing a normal command
on a computer over the transport had a bug in it that went unnoticed
because the code path was not tested. If the `stderr` of the command
contained any output the command would raise because the test that is
called `_computer_test_no_unexpected_output` would incorrectly return a
tuple of length one instead of two in that case.

In addition to adding tests to hit this code path, the message that is
printed in the case of non-empty stdout or stderr is deduplicated and
adapted to be bit clearer and refer directly to the documentation
instead of through a Github issue.