`verdi status`: distinguish database schema version incompatible (#4319)
If `verdi status` was called for a profile whose database schema version
is incompatible with the current code, a generic error was thrown that
no connection could be made to PostgreSQL. The connection is often fine,
it is just that AiiDA prohibits it until the database is made
compatible. Often one simply has to migrate the database after
installing a newer version of the code. This case is no caught
separately and the user is pointed to `verdi database migrate`.