Run the test install workflow with new pip dependency resolver (#4320)
As of version 20.2, `pip` ships with a new dependency solver, however,
since it is not yet ready for every-day-use, given that it is a lot more
strict in resolving dependency conflicts, it can only be enabled with
the flag `--use-feature=2020-resolver`. They plan to change this to the
default with v20.3 which will be released around October. To anticipate
this change, we already switch to use the new resolver in the normal
continuous integration workflow.

In addition, the test workflow is updated to use both the old and new
resolvers as well as testing installing without extras and all extras.