Add check to `verdi` to ensure extra dependencies are not imported (#4324)
The command `verdi devel check-undesired-imports` is added that checks
that when loading `verdi` there are no libraries imported that are part
of one of the extras and not of the base requirements. If this would be
the case `verdi` would except with normal execution. This command showed
that `seekpath` was imported in the `aiida.tools` top level module and
is now moved to prevent this.