Make the RabbitMQ connection parameters configurable
Up till now, the URL to connect to the RabbitMQ server was hardcoded.
This means it could only connect to the localhost over the standard port
and with the default credentials. Certain users, will require to deploy
RabbitMQ on a different machine than the AiiDA instance so the server
details should be configurable. Since this will no longer guarantee that
the RabbitMQ server is running on localhost, it should also be possible
to use SSL by changing the protocol from `amqp` to `amqps` and provide
specific user credentials.

The `aiida.manage.external.rmq.get_rmq_url` is responsible for
formatting the correct URI. The method takes the values that form the
scheme, netloc and path as arguments, whereas optional query parameters
can be specified through the keyword arguments. The supported arguments
are:

 * protocol
 * username
 * password
 * host
 * port
 * virtual_host

In addition, the following keyword arguments can be specified:

 * heartbeat  # heartbeat timeout in seconds
 * cafile  # string containing path to ca certificate file
 * capath  # string containing path to ca certificates
 * cadata  # base64 encoded ca certificate data
 * keyfile  # string containing path to key file
 * certfile  # string containing path to certificate file
 * no_verify_ssl  # boolean disables certificates validation should be "0" or "1"

Note that the hearbeat, unless explicitly specified, will be set to 600
seconds as a default.