slurm plugin: always raise for non-zero exit code
Prior to this change, cases where squeue would return a non-zero exit
code but an empty stderr would not lead to a SchedulerError.
This is fixed by adapting the slurm joblist command such that it is
always expected to produce exit code zero, and raising whenever a
non-zero exit code is encountered.