`Runner`: close loop when runner stops if runner created it (#4307)
If the loop is not closed, the file handles it managed during its
lifetime might not be properly cleaned up, leading to file leaks.
Therefore, if the loop is created by the `Runner` upon construction, it
should also close it when the runner closes. It cannot be done for loops
passed into the constructor, because they migth actually still be in
use by other parts of the code.