Make `--prepend-text` and `--append-text` options properly interactive (#4318)
The `--prepend-text` and `--append-text` options, used for both the
`verdi computer setup` and `verdi code setup` commands were normal
options as opposed to all other options that are interactive options.
The `InteractiveOption` is a custom option class type we developed that
will present the user with a prompt if it was not explicitly provided on
the command line. The `--non-interactive` flag can be used by the user
to prevent prompting, as long as a default value is defined.

The two options in question were implemented differently, presumably
because they take a potentially multiline string, which is not easily
defined on a prompt and therefore instead these options would be defined
through a text editor. The prompting of this text editor, if necessary,
was however not performed in the option itself, but in the command body.
At this point, the prompt cycle of the parameter cycle controlled by
click is already over.

Additionally, since the function checking whether the option had been
defined also considered an empty string as undefined, despite it being
the default, a specified empty string would still lead to the user to
be prompted, even when specifying `--non-interactive`.

The fix is to make both options proper interactive options just like the
rest. To this end, we create the `TemplateInteractiveOption` that works
just like the `InteractiveOption` with the only difference being that it
uses a file editor instead of an inline prompt. This change does force
us to have both options pop up their own file editor, whereas before
they were joined in a single file and both files were specified in the
same file, separated by a header that we defined.