`SlurmScheduler`: always raise for non-zero exit code (#4332)
The `SlurmScheduler` intentionally ignored non-zero exit codes returned
by SLURM when asking the status for a number of job ids. This was put in
place because SLURM will return a non-zero exit code not only in case of
actual errors in attempting to retrieve the status of the requested jobs
but also when specifying just a single job that no longer is active.

Since the latter is not really an error, yet is difficult to distinguish
from a "real" error, the exit code was ignored. However, this could
lead to the plugin sometimes incorrectly ignoring a real problem and
assuming a job was completed when it was in fact still active.

The solution is to use the weird behavior of SLURM that when asking for
more than one job, it will never return a non-zero status, even when
one or more jobs have finished. That is why, when asking for the status
of a single job, we duplicate the job id, such that even when it is no
longer active, the exit status will still be zero.