CI: limit upper version of `setuptools<50` for Jenkins build (#4343)
Builds on Jenkins started failing after `setuptools==0.50.0` was
released on August 30, 2020. This new version gets automatically
installed when using a `pyproject.toml` for the build system, but it
causes the installation of the `aiida-core` package through `pip` to
fail with the exception:

    ModuleNotFoundError: No module named 'setuptools._distutils'

Temporarily limiting the version of `setuptools` in the `pyproject.toml`
works around the problem for the time being.

The build is updated to also update the version of `pip` before
installing the package.