`Dict`: allow setting attributes through setitem and `AttributeManager` (#4351)
It was not possible to change the value of a key, either directly on the
node through setitem, nor through setattr via the `AttributeManager`
returned by the `dict` property, even though there is no reason not to
allow this, as long as the node is not stored. The changes in this commit
now allow the following pattern:

    node = Dict()
    node['x'] = 'Set a value for x'
    node.dict.x = 'Overwrite the value for x'

The `__setattr__` on the node itself is intentionally not implemented
because it would conflict with all the existing properties inherited
from the base classes.