Fix profile creation in the test fixture manager (#4360)
The test profile provided by the test fixture manager was broken after
the recent added feature where the broker configuration becomes
configurable. Since the test fixture uses custom code to create the
profile that is not tested in the test suite of `aiida-core`, it went
unnoticed that it was not updated to also include the broker
information. Adding the broker defaults fixes the problem, but really
the fixture code should be rewritten to not have its own profile
generation code.