`verdi export migrate`: add `--in-place` flag to migrate archive in place (#4220)
When an export archive needs to be migrated, one often does not care
about the original archive and simply wants to overwrite the existing
file with the migrated archive. The new flag `--in-place` saves users
from having to specify a temporary filename and copying it over the
original file after it has been migrated.

This commit also changes the exit code for migrating an export file that
already is up to date from an error to success (0).