`CalcJob`: improve logging in `parse_scheduler_output`
The level of the log that is fired if no detailed job info is available
is changed from `WARNING` to `INFO`. Since not all schedulers implement
the feature of retrieving this detailed job info, such as the often used
`DirectScheduler`, using a warning is not very apt. If the information
is missing, nothing is necessarily wrong, so `INFO` is better suited.

On the contrary, if the `Scheduler.parse_output` excepts, that is grave
and so its level is changed from a warning to an error.

Finally, a new condition is added where the scheduler does implement the
method to retrieve the detailed job info, but the command fails. In this
case, the return value will be non-zero. This value is now checked
explicitly and if the case, a info log is fired and the detailed job
info is set to `None`, which will cause the parsing to be skipped. This
case can for example arise when using the `SlurmScheduler` plugin, which
does implement the detailed job info feature, however, not all SLURM
installations have the job accounting feature enabled, which is required
by the plugin.