CI: continue on errors for install jobs using pip beta 2020 resolver (#4369)
Update the `test-install` workflow to not fail when any of the install
jobs with pip new dependency resolver fails. This new feature is in
development and so bugs are expected that are unrelated to our changes
and so its failure should not fail our builds which disrupts our own
CI process too much.

The compromise here is that in order to not block merges, the associated
steps and jobs still get a little green check-mark, and a simple warning
is displayed if issues with the 2020-resolver are encountered.