Depedencies: remove upper limit and allow `numpy~=1.17` (#4378)
The limit was introduced in `f5d6cba2baf0e7ca69b742f7e76d8a8bbcca85ae`
because of a broken pre-release. Now that a stable release is out, the
requirement is relax to allow newer versions as well. Note that we keep
the minimum requirement of `numpy==1.17` following AEP 003.

One change had to be applied in the code to make it compatible with newer
versions of `numpy`. In the legacy kpoints implementation, the entries
in `num_points` are of type `numpy.float64` for recent versions of
`numpy`, but need to be integers so they can be used for indexing in
`numpy.linspace()` calls.