Do not run `process_handler` when `exit_codes=[]`.
When a `process_handler` explicitly gets passed an empty
`exit_codes` list, it would previously always run. This is
now changed to _not_ run the handler instead.

The reason for this change is that it is more consistent with
the semantics of passing a lit of exit codes, where it only
triggers if the child process has any of the listed exit codes.