Prepare the code for the new repository implementation (#4344)
In `v2.0.0`, the new repository implementation will be shipped, that
despite our best efforts, requires some slight backwards-incompatible
changes to the interface. The envisioned changes are translated as
deprecation warnings:

 * `FileType`: `aiida.orm.utils.repository` ->`aiida.repository.common`
 * `File`: `aiida.orm.utils.repository` ->`aiida.repository.common`
 * `File`: changed from namedtuple to class
 * `File`: iteration is deprecated
 * `File`: `type` attribute -> `file_type`
 * `Node.put_object_from_tree`: `path` -> `filepath`
 * `Node.put_object_from_file`: `path` -> `filepath`
 * `Node.put_object_from_tree`: `key` -> `path`
 * `Node.put_object_from_file`: `key` -> `path`
 * `Node.put_object_from_filelike`: `key` -> `path`
 * `Node.get_object`: `key` -> `path`
 * `Node.get_object_content`: `key` -> `path`
 * `Node.open`: `key` -> `path`
 * `Node.list_objects`: `key` -> `path`
 * `Node.list_object_names`: `key` -> `path`
 * `SinglefileData.open`: `key` -> `path`
 * Deprecated use of `Node.open` without context manager
 * Deprecated any other mode than `r` and `rb` in the methods:
    o `Node.open`
    o `Node.get_object_content`
 * Deprecated `contents_only` in `put_object_from_tree`
 * Deprecated `force` argument in
    o `Node.put_object_from_tree`
    o `Node.put_object_from_file`
    o `Node.put_object_from_filelike`
    o `Node.delete_object`

The special case is the `Repository` class of the internal module
`aiida.orm.utils.repository`. Even though it is not part of the public
API, plugins may have been using it. To allow deprecation warnings to be
printed when the module or class is used, we move the content to a
mirror module `aiida.orm.utils._repository`, that is then used
internally, and the original module has the deprecation warning. This
way clients will see the warning if they use it, but use in `aiida-core`
will not trigger them. Since there won't be a replacement for this class
in the new implementation, it can also not be replaced or forwarded.