`Group`: add support for setting extras on groups (#4328)
The `DbGroup` database models get a new JSONB column `extras` which will
function just like the extras of nodes. They will allow setting mutable
extras as long as they are JSON-serializable.

The default is set to an empty dictionary that prevents the ORM from
having to deal with null values. In addition, this keeps in line with
the current design of other database models. Since the default is one
defined on the ORM and not the database schema, we also explicitly mark
the column as non-nullable. Otherwise it would be possible to still
store rows in the database with null values.

To add the functionality of setting, getting and deleting the extras to
the backend end frontend `Group` ORM classes, the corresponding mixin
classes are added. The functionality for the `BackendGroup` was already
accidentally added in a previous commit `65389f4958b9b111756450ea77e2`
so only the frontend is touched here.