`verdi setup`: forward broker defaults to interactive mode (#4405)
The options for the message broker configuration do define defaults,
however, the interactive clones for `verdi setup`, which are defined in
`aiida.cmdline.params.options.commands.setup` override the default with
the `contextual_default` which sets an empty default, unless it is taken
from an existing profile. The result is that for new profiles, the
broker options do not specify a default, even though for most usecases
the defaults will be required. After the changes of this commit, the
prompt of `verdi setup` will provide a default for all broker parameters
so most users will simply have to press enter each time.