REST API: list endpoints at base URL
The base URL of the REST API was returning a 404 invalid URL response
without providing any guidance to new users as to how to use the API.
We change this to return the list of endpoints formerly available only
under /server/endpoints.

Documentation of where to find the list of endpoints -- which seems
to have been entirely deleted -- is added.