Dependencies: update requirement `pytest~=6.0` and use `pyproject.toml` (#4410)
Starting from v6.0, `pytest` supports using the `pyproject.toml` instead
of a `pytest.ini` to define its configuration. Given that this is
quickly becoming the Python packaging standard and allows us to reduce
the number of configuration files in the top level of the repository, we
increase the version requirement of `pytest`.

Note that we also require `pytest-rerunfailures>=9.1.1` because lower
versions are broken in combination with `pytest==6.1`. See the following:

   https://github.com/pytest-dev/pytest-rerunfailures/issues/128

for details.