CI: add coverage patch threshold to prevent false positives (#4413)
The project diff percentage is the change in coverage w.r.t. all lines
in the project, whereas the patch diff percentage is the change in
coverage w.r.t. only lines touched by the PR. The patch threshold is
currently defaulting to 0%, hence it is very easy to fail. By raising it
to 0.1% it should now only fail when there is a significant reduction
in coverage. Number may need to be further tweaked.