`ProcessBuilder`: allow unsetting of inputs through attribute deletion (#4419)
The builder object was already able to delete set inputs through the
`__delitem__` method, but `__delattr__` was not implemented causing
`del builder.input_name` to raise. This is not consistent with how these
inputs can be set or accessed as both `__getattr__` and `__setattr__`
are implemented. Implementing `__delattr__` brings the implementation
up to par for all attribute methods.