Add `--delete-nodes` flag to verdi group delete
Remove the `--clear` flag for the `verdi group delete` command, in
favour of printing a warning that indicates the number of nodes in the
group before the prompt for user confirmation.

By default, the command only deletes the group, not the nodes it
contains. Add the `--delete-nodes` flag that allows the user to delete
all nodes in the group as well, according to standard traversal rules.

The `--force` flag removes the warnings and confirmation prompts for
both the group deletion, as well as deleting the nodes in the group
when the `--delete-nodes` flag is included.