Add `verdi group move-nodes` command
Add an extra subcommand `move-nodes` to `verdi group` that moves nodes
from a source group to a target group, instead of having to use the
`remove-nodes` and `add-nodes` subcommands in order to achieve this.

Similar to the `remove-nodes` and `add-nodes` commands, the move can be
made without prompts by adding the `--force` flag, and the nodes are
presented as a list of arguments.

Add corresponding test method to the `TestVerdiGroup` class of the
test_group.py suite.