Non-zero return code upon error in `verdi import`
`verdi import` will now exit with return code `1` if an error occurs at
any point during the import/migration.

A `--fail-fast` flag is added to `verdi import`, which will ensure an
immediate exit if an error occurs during the import/migration of a list
of archives (either local files or URLs).