node.open: more useful warnings
aiida-core started warning when methods like `Node.read` and `Node.del`
are used outside a context manager.
The warning, however, did not include any indication of where the
problematic function call was made (nor did it include instructions on
how to fix it).

Furthermore, this PR fixes an overlooked use of `Node.read` in
aiida-core that did not rely on the context manager.