♻️ Refactor archive exporter
This refactor fully decouples the extraction of data from the database, from the writing of that data to an archive.

It provides an abstract writer class, then each concrete writer is identified by a string (this could in fact be made into an entry point)

This code is backwards incompatible, in terms of the Python API, in that it removes the `extract_zip` and `export_tar` functions, but should still work with the verdi CLI, which only calls the `export` fcuntion, whose signature remains the same.

The commit also adds a backward compatible install requirement, `dataclasses`, which is included in python core from python 3.7 onwards.

The tests have not yet been altered to accomodate for these changes.