`CalcJob`: support nested directories in target of `remote_copy/symlink_list` (#4416)
The `upload_calculation` transport task would fail if either the
`remote_copy_list` or `remote_symlink_list` contained a target filepath
that had a nested directory that did not exist yet in the remote working
directory. Instead of inspecting the file system or creating the folders
remotely each time a nested target path is encountered, which would incur
a potentially expensive operation over the transport each time, the
directory hierarchy is first created in the local sandbox folder before
it is copied recursively to the remote in a single shot.