Docs: add "How to install plugins" section (#4468)
There was no centralized location yet that explains how a plugin package
is installed, yet this is one of the most critical first steps for new
AiiDA users. A new how-to section is created "How to install plugins"
that details this information. Since a file `plugins.rst` already
existed, which contains information on how to develop a plugin package,
it is renamed to `plugins_develop.rst`.