Docs: add the "How to explore the provenance graph" section
This section adds completely new documentation on the graph exploring
methods `get_incoming` and `get_outgoing` that were not yet documented.
It also documents the `creator`, `caller` and `creator` properties that
also weren't documented yet.

The `working_with_aiida/resultmanager.rst` is removed and its content,
that documents the `inputs` and `outputs` properties of process nodes as
well as the `res` property of the `CalcJobNode` class, is added to the
new section in heavily rewritten form.