Refactor archive progress bar (#4504)
This commit introduces a new generic progress reporter interface (in `aiida/common/progress_reporter.py`),
that can be used for adding progress reporting to any process.
It is intended to deprecate the existing `aiida/tools/importexport/common/progress_bar.py` module.

The reporter is designed to work similar to logging,
such that its "handler" is set external to the actual function, e.g. by the CLI.
Its default implementation is to do nothing (a null reporter),
and there is convenience function to set a [tqdm](https://tqdm.github.io/) progress bar implementation (`set_progress_bar_tqdm`).

The reporter is intended to always be used as context manager,
e.g. to allow the progress bar to be removed once the process is complete.

The reporter has been implemented in the archive export module,
and it is intended that it will also be implemented in the archive import module.
At this point the existing `aiida/tools/importexport/common/progress_bar.py` module can be removed.