Fix `UnboundLocalError` in `aiida.cmdline.utils.edit_multiline_template` (#4436)
If `click.edit` returns a falsy value, the following conditional would
be skipped and the `value` variable would be undefined causing an
`UnboundLocalError` to be raised. This bug was reported by @blokhin but
the exact conditions under which it occurred are not clear.

Cherry-pick: 861a39f268954833385e699b3acbd092ccd04e5e