`verdi migrate`: make `--in-place` work across different file systems (#4393)
The `verdi migrate` command assumed implicitly that the archive that is
to be migrated, resides on the same file system as the one that is used
by the `tempfile` module. If this is not the case, the `os.rename` call
used to atomically move the migrated archive to the original will fail
with the exception:

    OSError: [Errno 18] Invalid cross-device link

Changing `os.rename` to `shutil.move` fixes this problem. The downside,
however, is that the move is no longer atomic, but that is probably why
`os.rename` is restricted to same filysystem operations.