Add `reset` method to`ProgressReporterAbstract`
This commit adds the update method to the progress reporter.
This in-turn, allows for the implmentation of a callback mechanism
in `ArchiveReaderAbstract.iter_node_repos`.
The callback implmentation is taken from the
[disk-objectstore](https://github.com/aiidateam/disk-objectstore) package,
and so facilitates a later migration to its use.

The commit also moves the common repository import code
out of the backend specific modules, reducing code duplication.