Docs: Move data types to Topics section (#4469)
There was still content from the previous documentation on data types that was missing from the revamped docs. Here we move this content into the "Topics - Data types" section.

Also, restructured the content to first discuss the 'core' types, which we intend to keep in aiida-core in the long term, then the 'materials science' types, which may be moved to a plugin at some point.
Moved the export content to the end of the file.

Additionally, written a short text for each data type, which can start from the "Aim" bullet point for each type. Moved the other bullet points into a table, perhaps one for the "Core" and "Materials design" types each. Showed some simple use cases (e.g. how to load a dictionary into a Dict node) for each of the data types.