Make process functions submittable (#4539)
The limitation that process functions were not submittable, meaning they
could not be sent to a daemon worker but could only be run by the current
interpreter, was a historical one. Before the introduction of the system
of processes in v1.0, a `calcfunction` was nothing more than the
execution of a normal function. However, now, a process function creates
a `Process` instance in the background, just as any other process. This
means it can also be serialized and deserialized by a daemon worker.

Here we remove the limitation of process functions not being submittable
simply by removing the check. Note that there is no need to change the
implementation other than adding two attributes on the decorated function
that specify the corresponding process class and the method that allows
to recreate the instance from the serialized instance.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>