`verdi group delete`: deprecate and ignore the `--clear` option (#4357)
Note that the option is still there but no longer makes a difference. It
now merely prints a deprecation warning, but is otherwise ignored. The
reason is that otherwise, users would be forced to continue to use it
despite it raising a deprecation warning. The only danger is for users
that have come to depend on the slightly weird behavior that in order
to delete non-empty groups, one would have to pass them `--clear` option
otherwise the command would fail. After this change, this would now
delete the group without complaining, which may break this use case.
This use case was estimate to be unlikely and so it was accepted to
simply ignore the option.