SqlAlchemy: improve the alembic migration code
The code in `aiida.backends.sqlalchemy.manager` is used, among other
things, to control the database migrations. It can be used to get info
like the current revision of the database, current revision of the code
and to actually migrate the database. The migrations are performed
through `alembic`. This library has the somewhat weird approach of
loading the `env.py` file, which then actually calls the migrations.
This file somehow needs to get the configuration of the environment, for
example the connection to the database, and it does so by loading it
from the config that is imported. This config is actually built
dynamically in the `SqlBackendManager`.

The old implementation was correctly building up the configuration
object but was then calling the CLI commands of `alembic` to perform the
various operations, which was unnecessarily complex and inefficient. For
most operations, essentially everything but performing the actual
migrations, one can construct a `MigrationContext` directly, which
allows to inspect the state of the database, for example to get the
current database revision. For anything related to the revisions in the
code, e.g., to get the current head of the code, we can instantiate an
instance of the `ScriptDirectory` which allows one to inspect the
available revisions.

Through these changes, the `env.py` file is no longer loaded everytime,
even when it is not necessary, such as when simply retrieving the
current revision of the database.